pyproject.toml
[project]
name = "dscp"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "deepface>=0.0.99",
    "facenet-pytorch>=2.5.3",
    "kaggle>=2.0.0",
    "kagglehub>=1.0.0",
    "matplotlib>=3.10.8",
    "opencv-python>=4.13.0.92",
    "pandas>=3.0.2",
    "pillow>=12.2.0",
    "polars>=1.39.2",
    "pre-commit>=4.5.1",
    "requests>=2.32.5",
    "ruff>=0.15.7",
    "seaborn>=0.13.2",
    "tf-keras>=2.21.0",
    "torch>=2.11.0",
    "torchvision>=0.26.0",
    "tqdm>=4.67.3",
]

[tool.ruff]
line-length=79

[tool.ruff.lint]
select = ["E501"]
